refactor MoodTemplate parser
