+++ verbecc/src/defs/types/data/tense_template.py
@@ -1,5 +1,6 @@
 from typing import List
 
+from verbecc.src.defs.types.data.element import Element
 from verbecc.src.defs.types.data.person_ending import PersonEnding
 from verbecc.src.defs.types.lang_code import LangCodeISO639_1 as Lang
 from verbecc.src.defs.types.mood import Mood
@@ -7,7 +8,7 @@ from verbecc.src.defs.types.tense import Tense
 from verbecc.src.defs.types.person import Person
 
 
-class TenseTemplate:
+class TenseTemplate(Element):
 
     def __init__(
         self, lang: Lang, mood: Mood, tense: Tense, person_endings: List[PersonEnding]
